backend/foodgram/requirements.txt
django==3.2.15 
django-filter==22.1 
djangorestframework==3.14.0 
djangorestframework-simplejwt==4.7.2 
djoser==2.1.0 
Pillow==9.2.0 
psycopg2-binary==2.8.6 
PyJWT==2.5.0 
python-dotenv==0.21.0 
pytz==2022.2.1 
sqlparse==0.4.3 
requests==2.26.0
asgiref
gunicorn==20.0.4